requirements.txt
# requirements.txt
pillow
beautifulsoup4
lxml
